+++ src/data/index.js
@@ -1,11 +1,15 @@
 import courses from "./Courses";
 import digitalTwin from "./digitalTwin";
 import jobs from "./Jobs";
+import SkillsUtils from "./skills.utils";
 
 const data = {
     courses,
     jobs,
     digitalTwin,
+    utils: { 
+        skills : SkillsUtils,
+    },
 }
 
 export default data;